# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "面向生产环境的本地文件扫描、持久化上传队列与 rclone 调度 API。",
        "title": "Rclone Sync Hub API",
        "contact": {
            "name": "Project repository",
            "url": "https://github.com/zane626/rclone-sync-hub-go-backend"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/watch-folders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按状态分页获取监听文件夹列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-folders"
                ],
                "summary": "列出监听文件夹",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态: detecting|watching|stopped|paused|error",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词：对 name/local_path/remote_name/remote_path 模糊查询",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从 1 开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "items 为列表，total 为总数",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "新增一个被监听并同步到指定 rclone remote 的本地目录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-folders"
                ],
                "summary": "创建监听文件夹",
                "parameters": [
                    {
                        "description": "监听文件夹配置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.WatchFolderCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rclone-sync-hub_internal_model.WatchFolder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/watch-folders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-folders"
                ],
                "summary": "获取监听文件夹详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "监听文件夹 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rclone-sync-hub_internal_model.WatchFolder"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-folders"
                ],
                "summary": "更新监听文件夹",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "监听文件夹 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容（全部可选）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.WatchFolderUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rclone-sync-hub_internal_model.WatchFolder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "原子取消该目录的未完成任务并解除文件快照归属；活动扫描期间返回冲突",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-folders"
                ],
                "summary": "删除监听文件夹",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "监听文件夹 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_api.WatchFolderCreateReq": {
            "type": "object",
            "required": [
                "local_path",
                "name",
                "remote_name",
                "remote_path"
            ],
            "properties": {
                "filter_keywords": {
                    "description": "可选，多行关键字，换行分隔，扫描时模糊匹配排除",
                    "type": "string"
                },
                "local_path": {
                    "type": "string"
                },
                "max_depth": {
                    "description": "可选，0 表示不限制",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remote_name": {
                    "type": "string"
                },
                "remote_path": {
                    "type": "string"
                },
                "scan_interval_seconds": {
                    "description": "可选，默认 300",
                    "type": "integer"
                },
                "sync_type": {
                    "description": "可选，默认 local_to_remote",
                    "type": "string"
                }
            }
        },
        "internal_api.WatchFolderUpdateReq": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "filter_keywords": {
                    "type": "string"
                },
                "local_path": {
                    "type": "string"
                },
                "max_depth": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remote_name": {
                    "type": "string"
                },
                "remote_path": {
                    "type": "string"
                },
                "scan_interval_seconds": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "sync_type": {
                    "type": "string"
                }
            }
        },
...
        "rclone-sync-hub_internal_model.WatchFolder": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "enabled": {
                    "description": "是否启用该监听",
                    "type": "boolean"
                },
                "failedFileCount": {
                    "description": "累计上传失败文件数",
                    "type": "integer"
                },
                "filterKeywords": {
                    "description": "过滤关键字，多行存储，每行一个；路径或文件名包含任一关键字（模糊匹配）则排除",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastActiveAt": {
                    "description": "最近有文件变更 / 上传的时间",
                    "type": "string"
                },
                "lastError": {
                    "description": "最近一次错误信息（便于排查）",
                    "type": "string"
                },
                "lastScanAt": {
                    "description": "兼容字段：最近一次扫描开始时间",
                    "type": "string"
                },
                "lastScanDurationMs": {
                    "description": "最近一次扫描耗时（毫秒）",
                    "type": "integer"
                },
                "lastScanFinishedAt": {
                    "description": "最近一次扫描结束时间（无论成功失败）",
                    "type": "string"
                },
                "lastScanStartedAt": {
                    "description": "最近一次扫描开始时间",
                    "type": "string"
                },
                "lastScanSuccessAt": {
                    "description": "最近一次成功扫描结束时间",
                    "type": "string"
                },
                "lastSyncAt": {
                    "description": "最近一次同步时间",
                    "type": "string"
                },
                "localPath": {
                    "description": "注意：MySQL InnoDB 单列索引最大 3072 字节（utf8mb4 约 768 字符），且 LocalPath 上有唯一索引，因此长度限制为 768。",
                    "type": "string"
                },
                "maxDepth": {
                    "description": "最大监听深度，0 表示不限制",
                    "type": "integer"
                },
                "name": {
                    "description": "基本配置",
                    "type": "string"
                },
                "nextScanAt": {
                    "description": "预计下一次扫描时间（可选）",
                    "type": "string"
                },
                "remoteName": {
                    "description": "rclone remote 名称",
                    "type": "string"
                },
                "remotePath": {
                    "description": "远端路径",
                    "type": "string"
                },
                "scanIntervalSeconds": {
                    "description": "配置相关扩展",
                    "type": "integer"
                },
                "scanLeaseExpiresAt": {
                    "type": "string"
                },
                "scanLeaseOwner": {
                    "type": "string"
                },
                "status": {
                    "description": "状态信息",
                    "type": "string"
                },
                "syncType": {
                    "description": "同步类型",
                    "type": "string"
                },
                "totalFileCount": {
                    "description": "统计维度（便于分析与展示）",
                    "type": "integer"
                },
                "totalFileSize": {
                    "description": "当前已知文件总大小（字节）",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uploadedBytes": {
                    "description": "累计已上传字节数",
                    "type": "integer"
                },
                "uploadedFileCount": {
                    "description": "累计已成功上传文件数",
                    "type": "integer"
                },
                "windowUploadedBytes": {
                    "description": "窗口内上传字节数",
                    "type": "integer"
                },
                "windowUploadedFiles": {
                    "description": "最近一段时间的窗口统计（可用于展示趋势）",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}